# api/openapi.yaml
openapi: 3.1.0
info:
  title: ame API
  description: Adaptive Mastery Engine — quiz & exam platform with agent identity.
  license:
    name: ''
  version: 0.1.0
paths:
...
  /v1/auth/login:
    post:
      tags:
      - auth
      summary: POST /v1/auth/login — authenticate with email and password.
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginBody'
        required: true
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
...
  /v1/auth/register:
    post:
      tags:
      - auth
      summary: POST /v1/auth/register — register a new user with email and password.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterBody'
        required: true
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          description: Validation failed
...
components:
  schemas:
...
    AuthResponse:
      type: object
      required:
      - token
      - user
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/UserInfo'
...
    LoginBody:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
...
    RegisterBody:
      type: object
      required:
      - email
      - name
      - password
      - role
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
...
    UserInfo:
      type: object
      required:
      - id
      - name
      - email
      - role
      properties:
        email:
          type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
...